Fix clang warning about comparing char with -1.
We store in the char variable a number of found attributes and -1 when
we found everything we need (see VT_TOKEN_ATTRS_MAX define). It seems that
we store max 2 attributes at all.

I changed type of this field from char to signed short.

+++ src/drivers/char/terminal.c
@@ -90,7 +90,7 @@ static struct vt_token *vt_from_term_token(struct vt_token *vt_token,
 	char char2 = DECODE_CHAR2(terminal_token);
 	char code = DECODE_CODE(terminal_token);
 
-	int attrs_len;
+	signed short attrs_len;
 
 	if (!is_valid_action(action)) {
 		return NULL;
